[SCCP] get rid of potentially dangling iterator
getUserBonus took an iterator into KnownConstants, but then inserted
into it later in the function, which could resize the data-structure
(DenseMap is based on a vector, so is not iterator stable).

The iterator is not needed, because we never change the value of an
entry after inserting.

Pull Request: https://github.com/llvm/llvm-project/pull/105609